git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/branches/b_djl_cmd_line_parser_fixes@572 41062d70-b26d-460a-9302-f438a8930db8

+++ src/TinyCadMultiDoc.cpp
@@ -187,7 +187,7 @@ void CTinyCadMultiDoc::InsertSheet(int i, CTinyCadDoc *pDoc)
 	{
 		m_sheets.insert(it, pDoc);
 	}
-	else if (i != m_sheets.size())
+	else if (i != static_cast<int>(m_sheets.size()))
 	{
 		m_sheets.insert(it + i + 1, pDoc);
 	}
